Feat: GeneralLayout 구현 (#96)
- 모든 페이지에 적용될 컴포넌트
- 탭바를 포함
- 사용자 인증 로직을 App에서 옮김
